+++ api/controllers/artist.contoller.js
@@ -1,7 +1,5 @@
 import prisma from '../lib/prisma.js';
 
-// TODO: CONTROL IF USER IS ADMIN AND VERIFY ADMIN TOKEN -> DO NOT FORGET TO MODIFY TOKEN FOR IS_ADMIN
-
 export const getArtist = async (req, res) => {
   const id = req.params.id;
 
@@ -44,8 +42,6 @@ export const getArtists = async (req, res) => {
 };
 
 export const addArtist = async (req, res) => {
-  // TODO: modify adding authorization accessible only for admins
-
   const body = req.body;
 
   try {
@@ -62,19 +58,26 @@ export const addArtist = async (req, res) => {
 };
 
 export const updateArtist = async (req, res) => {
-  // TODO: modify updating authorization accessible only for admins
-
   const body = req.body;
   const id = req.params.id;
+  const tokenUserId = req.userId;
+
   try {
-    const artist = await prisma.artist.update({
+    const artist = await prisma.artist.findUnique({
+      where: { id },
+    });
+
+    if (artist.userId !== tokenUserId)
+      return res.status(403).json({ message: 'Forbidden! Not authorized!' });
+
+    const updatedArtist = await prisma.artist.update({
       where: { id },
       data: {
         ...body,
       },
     });
 
-    res.status(200).json(artist);
+    res.status(200).json(updatedArtist);
   } catch (error) {
     console.log(error);
     res.status(500).json({ message: 'Failed to update artist info!' });
@@ -82,11 +85,17 @@ export const updateArtist = async (req, res) => {
 };
 
 export const deleteArtist = async (req, res) => {
-  // TODO: modify deleting authorization accessible only for admins
-
   const id = req.params.id;
+  const tokenUserId = req.userId;
 
   try {
+    const artist = await prisma.artist.findUnique({
+      where: { id },
+    });
+
+    if (artist.userId !== tokenUserId)
+      return res.status(403).json({ message: 'Forbidden! Not authorized!' });
+
     await prisma.artist.delete({
       where: { id },
     });
